feat: Authorization check

+++ api/controllers/feed.js
@@ -108,6 +108,11 @@ function updatePost(req, res, next) {
       error.statusCode = 404;
       throw error;
     }
+    if (post.creator.toString() !== req.userId) {
+      const error = new Error('Not authorized');
+      error.statusCode = 403;
+      throw error;
+    }
     if (imageUrl !== post.imageUrl) {
       clearImage(post.imageUrl);
     }
@@ -132,11 +137,23 @@ function deletePost(req, res, next) {
       throw error;
     }
 
+    if (post.creator.toString() !== req.userId) {
+      const error = new Error('Not authorized');
+      error.statusCode = 403;
+      throw error;
+    }
+
     clearImage(post.imageUrl);
     return Post.findByIdAndRemove(postId);
   })
   .then(result => {
-    console.log(result);
+    return User.findById(req.userId);}
+  )
+  .then(user => {
+    user.posts.pull(postId);
+    return user.save();
+  })
+  .then(result => {
     res.status(200).json({message: 'Post deleted succesfully.'});
   })
   .catch(catchError);
